pyproject.toml
[build-system]
requires = ["hatchling>=1.24"]
build-backend = "hatchling.build"

[project]
name = "endopigraph-ajmorph"
version = "0.1.0"
description = "EndoPiGraph-AJmorph v1: build endothelial pi-graphs and adherens-junction morphology features from microscopy images."
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}
authors = [
  {name = "EndoPiGraph Team"}
]
dependencies = [
  "numpy>=1.23",
  "pandas>=1.5",
  "scikit-image>=0.21",
  "scipy>=1.10",
  "tifffile>=2023.7.10",
  "networkx>=3.1",
  "matplotlib>=3.7",
  "PyYAML>=6.0",
  "tqdm>=4.66",
  "requests>=2.31"
]

[project.optional-dependencies]
cellpose = [
  "cellpose>=3.0.0"
]
ml = [
  "scikit-learn>=1.3"
]
dev = [
  "pytest>=7.4",
  "ruff>=0.5.0"
]

[project.scripts]
endopigraph = "endopigraph.cli:main"

[tool.hatch.build.targets.wheel]
packages = ["src/endopigraph"]
